fix: Wrap BottomBar contents in SafeArea
Wraps the contents of the BottomBar in a SafeArea. This adds a bit of
bottom padding to keep BottomBar contents outside of iOS safe area
insets on iPhone X and later devices.

Previously, the "How it's made" link was inside the iOS safe areas,
making it difficult to tap on iPhones.